+++ PointerChainTester/Program.cs
@@ -26,14 +26,18 @@ namespace au.edu.federation.PointerChainTester
         public static int TEXT_COMPARISON_CHAR_LIMIT = 33;
 
         public static bool connectedToProcess = false; // Keep track of whether we're connected to the named process or not
-        public static bool validPointerTrail = false; // Keep track of whether the pointer trail is legal or not
+        public static bool validPointerTrail  = false; // Keep track of whether the pointer trail is legal or not
 
-        public static string processName;      // The name of the process we want to attach to
-        public static Process[] processArray;  // Used to find the process
-        public static Process gameProcess;     // The actual process we found
-        public static int processHandle;       // The handle to the process we found
-        public static int processBaseAddress;  // The base address of the process we found
-        public static int featureAddress;      // The feature address after following the pointer trail
+        public static string processName;        // The name of the process we want to attach to
+        public static Process[] processArray;    // Used to find the process
+        public static Process gameProcess;       // The actual process we found
+
+        // Note: While integers in .NET are always 32-bit, IntPtr can be used to be flexible with the build platform - so when building
+        //       as an x86 app an IntPtr is 32-bit but when building as an x64 app an IntPtr is 64-bits.
+
+        public static IntPtr processHandle;      // The handle to the process we found
+        public static IntPtr processBaseAddress; // The base address of the process we found
+        public static IntPtr featureAddress;     // The feature address after following the pointer trail
 
         // The pointer trail to find our feature address as a list of strings
         public static List<string> pointerList = new List<string>();
